power up spawner
extension of the paddle script, it has a transform that is spawned onto with a amanger script that coordinates when and how things should spawn. a few caveats:
To reduce coupling the spawn object has a trigger volume that needs to be MANUALLY made parallel to the size of a given power up.

The spawn manager is an extension of an older system (and allows for some flexibility this way) where it gets the spawn point with a tag search. At this point, only the spawn transform has this tag BUT it has it as the "physics" object. Keep it that way.